refactor: extract workspace git helpers
